requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
optuna
joblib
dvc
pyyaml
python-dotenv
pydantic